Change price input to float for ticket sales

+++ recuperação/#1.py
@@ -1,5 +1,5 @@
 f = input('Digite o filme: ')
-p = int (input ('coloque o preço unitario:R$ '))
+p = float (input ('coloque o preço unitario:R$ '))
 q = int (input('quantos ingressos foi vendidos: '))
 
 v_total =  (p) * (q)
